+++ CHANGELOG.md
@@ -28,6 +28,12 @@ A preview of the next release can be installed from
 - Fix `resolve` when `*ns*` is bound to symbol
 - Bump `deps.clj` to `1.12.1.1550`
 - Bump `http-client` to `0.4.23`
+- Add classes for compatibility with [promesa](https://github.com/funcool/promesa):
+  - `java.lang.Thread$Builder$OfPlatform`
+  - `java.util.concurrent.ForkJoinPool`
+  - `java.util.concurrent.ForkJoinPool$ForkJoinWorkerThreadFactory`
+  - `java.util.concurrent.ForkJoinWorkerThread`
+  - `java.util.concurrent.SynchronousQueue`
 
 ## 1.12.200 (2025-04-26)